try fix

+++ 0x1A-hash_tables/2-key_index.c
@@ -8,7 +8,7 @@
  */
 unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
-	int mask = 511;
+	unsigned long int mask = 511;
 
 	return (hash_djb2(key) & mask);
 }
